Add support for jumps by splitting the CPU into three stages
Due to the propagation delay of our combinatorial logic (such as the
ALU), it is impossible to support jumps if every CPU instruction is only
given a single clock cycle.

In addition, reading a new instruction from ROM requires the CPU to wait
a cycle before the output value is valid: this fixed delay was not a
problem when instructions were executed one after the other, but the
presence of jumps means that ROM access and CPU operations have to be
precisely synchronised, to avoid accidentally executing the instruction
immediately following a jump that has been taken.

Because of this, the CPU has been split into three stages: instruction
fetch (IF), instruction decode (ID), and "write back" (WB). Note that,
when executing an A-instruction, the CPU is able to skip the WB stage,
as the ALU is not used and the instruction's immediate value can be
stored in the A register as part of the ID stage. On the other hand,
C-instructions require all three stages: during ID, the instruction's
operands are loaded into the ALU, and during WB, the ALU's output (which
has had enough time to propagate) is written to the destination
registers and, if required, the program counter is updated to jump to a
new location.