Add a rudimentary Keyboard driven by the three-button snap-off PMOD
According to the Hack specification, reading the contents of memory
address 0x6000 must return a 16-bit ASCII value representing the key
that is currently being pressed.

We don't yet support a full hardware keyboard, so for the time being we
use the snap-off PMOD on the iCEBreaker, which comes with three buttons,
to simulate left arrow, escape, and right arrow. Although extremely
limited, this particular set of keys would be enough for a game of Pong.