Implement VGA screen output backed by pseudo dual-port VRAM
This commit introduces a pseudo dual-port VRAM that is shared between a
VGA output module and the Hack CPU. The VRAM uses the lowest 16KiB of a
32KiB SPRAM chip and can be accessed at memory addresses 0x4000-0x5FFF.

The VGA screen has priority, read-only access to the shared VRAM. A
"read enable" signal is used to request a memory read, which happens
once every 16 pixels while in the active rendering portion of the
screen. Based on our 25.125MHz pixel clock, we can render a 640x480
picture at 60Hz; however, in accordance with the Hack specification, we
only display a 512x256 black-and-white "letterbox", with pixels falling
outside of that rectangle set to display the colour black.

When the VGA module is not reading from the VRAM, the "busy" signal goes
low and the CPU can request a read or write. Due to the shared VRAM
logic, each read or write incurs a two-cycle delay, which is handled by
a new mem_read stage in the CPU pipeline. In addition, memory reads are
cached in a new 16-bit register, which is updated by every instruction
that modifies the A register: specifically, all A-instructions as well
as C-instructions with the "d1" flag high. A further mem_fetch CPU state
is used to transfer data from the memory's output bus to the register.

This commit, along with the three-button keyboard introduced by 1472ada,
is sufficient to play a full game of Pong using the source code supplied
with the Nand2Tetris course.