Implement VGA screen output backed by pseudo dual-port VRAM
This commit introduces a pseudo dual-port VRAM that is shared between a
VGA output module and the Hack CPU. The VRAM uses the lowest 16KiB of a
32KiB SPRAM chip and can be accessed at memory addresses 0x4000-0x5FFF.

The VGA screen has priority, read-only access to the shared VRAM. A
"read enable" signal is used to request a memory read, which happens
every 16 pixels while in the active rendering portion of the screen.
Based on our 25.125MHz pixel clock, we can render a 640x480 picture at
60Hz. However, in accordance with the Hack specification, only a 512x256
black-and-white "letterbox" can be displayed, with pixels falling
outside of that rectangle set to display the colour black.

When the VGA module is not reading from the VRAM, the "busy" signal goes
low and the CPU can request a read or write. Due to the shared VRAM
logic, each read or write incurs a two-cycle delay, which is handled by
two additional stages in the CPU pipeline, mem_wait and mem_read. These
states are only reachable by C-instructions that use a memory operand
("a" flag high). Therefore, when executing either A-instructions, or
C-instructions that only use the A and C registers as operands, no delay
is incurred.

This commit, along with the three-button keyboard introduced by 7f76dd2,
is sufficient to play a full game of Pong using the source code supplied
with the Nand2Tetris course.