Simplify the reset signals used for the ROM and CPU
Instead of using the "locked" output of the PLL to reset the ROM, we use
it as a clock enable (clken) signal. This means that the CPU becomes the
only module that requires a reset signal, which is a combination of the
ROM's ready output signal and the physical reset provided by one of the
buttons on the iCEBreaker (BTN_N).