Speed up the CPU by speculatively fetching the next instruction
In a typical Hack program, jump instructions are relatively rare. For
example, only 3.2% of all instructions in the Pong sample code are jumps
(4.9% of all C-instructions).

Therefore, in the vast majority of cases, we can skip a CPU cycle (in
the inst_fetch state) by speculatively incrementing the program counter
before the current instruction has finished executing. If we're right,
the next instruction will be ready by the time we're finished with the
current one; and if we're wrong, we'll simply overwrite the speculative
program counter with the jump target and wait for an inst_fetch cycle.