Speed up the CPU by only accessing memory when necessary
Previously, we would speculatively read from memory whenever the value
of the A register changed. This made it easier to process C-instructions
with a memory operand, but it came with the added cost of having to
waste the occasional memory read cycle when the A register would change
but the corresponding memory address would never be used by the program.

This commit rearranges the CPU's state machine so that A-instructions
are dealt with in only two cycles (including instruction fetch), whereas
C-instructions take a variable amount of cycles depending on whether the
contents of memory need to be fetched and loaded into the ALU.

The only potential downside of this approach is that if two subsequent
C-instructions wish to use the same memory operand, then we may
unnecessarily end up reading it multiple times. However, given the way a
typical Hack program is structured, and considering that the ALU output
can be written back to multiple destinations at once (A, D and M), this
is an extremely unlikely scenario that is not worth optimising.