Replace the 3-button PMOD with a full PS/2 keyboard
Although the 3-button snap-off PMOD was enough to play a game of Pong,
the Hack specification expects a standard, full-sized keyboard that can
output ASCII characters as well as the standard control keys such as
arrow keys, function keys, return, etc.

This commit replaces the 3-button keyboard with one that interfaces with
a PS/2 driver. Only the keys required for the Hack specification are
implemented, along with full support for Caps Lock and both the left and
right Shift keys. The PS/2 driver is deliberately minimal and does not
support the full range of PS/2 keyboard functions and commands.

When the keyboard is first plugged in, the driver will send an "Enable
scanning" (0xF4) command, which will cause the keyboard to start sending
key presses over the PS/2 connection as they happen. The driver supports
decoding both one- and two-byte (extended) scan codes.

When a key is held down for a period of time, the keyboard will repeat
the same "key down" message roughly every 100 milliseconds. When the key
is released, a "key up" message will be sent including the scan code of
the key that is no longer pressed. As the Hack specification only
expects a single key to be pressed at any one time, in the case of
multiple key presses, this implementation always favours the most recent
one.

The only exception to the above is the way the Shift keys are handled,
as they must be tracked even as other keys are pressed, so the PS/2
driver will actively listen for "key up" events if they refer to a Shift
key that was previously pressed.

Caps Lock is handled slightly differently, as each press is meant to
toggle it on and off, and if the Caps Lock key is held down, it should
have no other effect. Whenever the Caps Lock changes state, the PS/2
driver will send a "Set LEDs" (0xED) command to the keyboard, asking it
to update the keyboard's own LED reflecting whether Caps Lock is on or
off. The other LEDs (Scroll Lock and Number Lock) are ignored and always
set to the off state.